\name{nocornacod-package}
\alias{nocornacod-package}
\alias{nocornacod}
\docType{package}
\title{\packageTitle{nocornacod}}
\description{\packageDescription{nocornacod}}
\section{Package Content}{\packageIndices{nocornacod}}
\author{\packageAuthor{nocornacod}}
\section{Maintainer}{\packageMaintainer{nocornacod}}
\keyword{package}
