% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ncrna.R
\name{predict_ncrna}
\alias{predict_ncrna}
\title{Predict ncRNA around a CDS.}
\usage{
predict_ncrna(
  fasta_path,
  coords_path,
  outdir_path,
  inner_pad = 100L,
  outer_pad = 11L,
  min_ncrna_len = 13L,
  min_score = 0.5,
  smooth_win = 27L,
  max_cons_fold = 79L,
  threads = 1L,
  force = FALSE
)
}
\arguments{
\item{fasta_path}{String. Path to a FASTA file of DNA sequences. Each
sequence should contain a CDS and additional flanking bases. Each
sequence should also have a unique ID where the first whitespace delimited
word after the '>', with any leading whitespace trimmed, is considered the
ID.}

\item{coords_path}{String. Path to a tab-delimited file with four columns
and one row for each sequence in \code{fasta_path}. The columns should be
sequence ID, start coordinate of CDS, end coordinate of CDS, and
strand of CDS (given as "+" or "-"). The file should not have a header.}

\item{outdir_path}{String. Path to the output directory. The function will
not overwrite an existing directory unless \code{force} is \code{TRUE}.}

\item{inner_pad}{Integerish scalar. Number of bases into the CDS to pad the
putative ncRNA region when predicting secondary structure.}

\item{outer_pad}{Integerish scalar. Number of bases outside the conserved
region to pad the putative ncRNA region when predicting covariance and
secondary structure.}

\item{min_ncrna_len}{Integerish scalar. Minimum number of bases past the
ends of the CDS that should be conserved for that region to be
considered a ncRNA.}

\item{min_score}{Double scalar. Minimum conservation scores for the ncRNA
in its multiple sequence alignment.}

\item{smooth_win}{Integerish scalar. Window size used when smoothing
conservation scores.}

\item{max_cons_fold}{Integerish scalar. Maximum number of sequences to use
when predicting covariance and secondary structure. This is just used to
speed up R-scape.}

\item{threads}{Integerish scalar. Number of threads to use. This is only
used in the alignment step.}

\item{force}{Force overwrite of output directory?}
}
\description{
There are some classes of non-coding RNA that are in very close proximity,
in DNA space, to the proteins they interact with. This function attempts to
predict these types of ncRNA by using DNA multiple sequence alignments of
clusters of highly similar proteins.
}
\details{
The pipeline begins by constructing a DNA multiple sequence alignment. Then
it calculates a per-column conservation score for the entire alignment.
Next it predicts change points in conservation scores. The pipeline operates
under the assumption that the ncRNAs of interest extend outside the CDS and
thus if there is some conserved ncRNA shared among the sequences, there will
also be sequence conservation outside of the CDS region. Therefore, it is
expected that the change point in conservation scores will be outside the
CDS region as well. As the final step, the pipeline computes covariation
scores and consensus folds for the predicted ncRNA region.
}
